reused bearer token try from impl in rocket guard
